requirements.txt
Flask~=1.1.2
msal~=1.10.0
PyYAML~=5.4.1
python-dotenv~=0.15.0
pymongo~=3.11.3